Fix MCP run-as CI size gate
